Move BridgeUtils.getAmountSentToAddress to BridgeUtilsLegacy
