+++ rskj-core/src/main/java/co/rsk/peg/BridgeUtilsLegacy.java
@@ -1,9 +1,12 @@
 package co.rsk.peg;
 
-import co.rsk.bitcoinj.core.Address;
-import co.rsk.bitcoinj.core.NetworkParameters;
+import co.rsk.bitcoinj.core.*;
+import co.rsk.config.BridgeConstants;
+import co.rsk.core.RskAddress;
+import co.rsk.crypto.Keccak256;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.config.blockchain.upgrades.ConsensusRule;
+import org.ethereum.core.Transaction;
 
 /**
  * @deprecated Methods included in this class are to be used only prior to the latest HF activation
@@ -68,5 +71,24 @@ public class BridgeUtilsLegacy {
         System.arraycopy(addressBytes, 1, hashBytes, 0, 20);
 
         return new Address(networkParameters, version, hashBytes);
+    }
+
+    /**
+     * Legacy version for getting the amount sent to a btc address. This is used in {@link BridgeSupport#registerFastBridgeBtcTransaction(Transaction, byte[], int, byte[], Keccak256, Address, RskAddress, Address, boolean)}
+     *
+     * @param constants
+     * @param btcTx
+     * @param btcAddress
+     * @return total amount sent to the given address.
+     */
+    @Deprecated
+    protected static Coin getAmountSentToAddress(BridgeConstants constants, BtcTransaction btcTx, Address btcAddress) {
+        Coin v = Coin.ZERO;
+        for (TransactionOutput o : btcTx.getOutputs()) {
+            if (o.getScriptPubKey().getToAddress(constants.getBtcParams()).equals(btcAddress)) {
+                v = v.add(o.getValue());
+            }
+        }
+        return v;
     }
 }
