+++ rskj-core/src/test/java/co/rsk/peg/BridgeUtilsLegacyTest.java
@@ -2,6 +2,8 @@ package co.rsk.peg;
 
 import co.rsk.bitcoinj.core.Address;
 import co.rsk.bitcoinj.core.BtcECKey;
+import co.rsk.bitcoinj.core.BtcTransaction;
+import co.rsk.bitcoinj.core.Coin;
 import co.rsk.config.BridgeConstants;
 import co.rsk.config.BridgeMainNetConstants;
 import co.rsk.config.BridgeRegTestConstants;
@@ -236,6 +238,36 @@ public class BridgeUtilsLegacyTest {
             activations,
             addressBytes
         );
+    }
+
+    public void testGetAmountSentToAddress(BridgeConstants constants, Coin valueToTransfer, Boolean includeOutput) {
+        Address receiver = constants.getGenesisFederation().getAddress();
+        BtcTransaction btcTx = new BtcTransaction(constants.getBtcParams());
+        if (includeOutput){
+            btcTx.addOutput(valueToTransfer, receiver);
+        }
+        Assert.assertEquals(valueToTransfer, BridgeUtilsLegacy.getAmountSentToAddress(constants, btcTx, receiver));
+    }
+
+    @Test
+    public void getAmountSentToAddress_coin() {
+        Coin valueToTransfer = Coin.COIN;
+        testGetAmountSentToAddress(bridgeConstantsRegtest, valueToTransfer, true);
+        testGetAmountSentToAddress(bridgeConstantsMainnet, valueToTransfer, true);
+    }
+
+    @Test
+    public void getAmountSentToAddress_no_output_for_address() {
+        Coin valueToTransfer = Coin.ZERO;
+        testGetAmountSentToAddress(bridgeConstantsRegtest, valueToTransfer, false);
+        testGetAmountSentToAddress(bridgeConstantsMainnet, valueToTransfer, false);
+    }
+
+    @Test
+    public void getAmountSentToAddress_output_value_is_0() {
+        Coin valueToTransfer = Coin.ZERO;
+        testGetAmountSentToAddress(bridgeConstantsRegtest, valueToTransfer, true);
+        testGetAmountSentToAddress(bridgeConstantsMainnet, valueToTransfer, true);
     }
 
     @Test
